# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI_Chatbot API Skeleton
  version: 0.1.0-skeleton
  description: Auto-generated skeleton from workbook; schemas are intentionally coarse.
servers:
- url: http://localhost:8080
paths:
...
  /v1/internal/events/ingest:
    post:
      operationId: opsEventIngest
      summary: 이벤트 수집
      description: 이벤트 수집 기능 제공
      tags:
      - 운영 관측 통합
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Workbook skeleton response (202)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: OPS-EVENT-INGEST
      x-source-row: 55
...
  /v1/internal/tools/execute:
    post:
      operationId: toolCallExecute
      summary: Tool 실행
      description: Tool 실행 기능 제공
      tags:
      - Tool MCP 연동
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: TOOL-CALL-EXECUTE
      x-source-row: 47
  /v1/internal/tools/validate:
    post:
      operationId: toolCallValidate
      summary: Tool 파라미터 검증
      description: Tool 파라미터 검증 기능 제공
      tags:
      - Tool MCP 연동
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Workbook skeleton response (201)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      x-access-level: AUTHENTICATED
      x-rbac-roles:
      - SYSTEM
      x-program-id: TOOL-CALL-VALIDATE
      x-source-row: 48
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT